refactor: simplify virtual export
