+++ packages/demo/src/server/index.ts
@@ -1,6 +1,6 @@
-import indexHtmlMiddleware from "virtual:@hiogawa/vite-index-html-middleware/hattip";
 import globApiRoutes from "virtual:glob-api-routes/hattip";
 import { compose } from "@hattip/compose";
+import { indexHtmlMiddleware } from "@hiogawa/vite-index-html-middleware/runtime";
 
 export function createHattipApp() {
   return compose(globApiRoutes(), indexHtmlMiddleware());
