+++ packages/vite-index-html-middleware/src/index.ts
@@ -1,13 +1,15 @@
 import type { Plugin } from "vite";
 
-// TODO: how to expose as non virtual? (so that it's easy to provide import typing?)
-const virtualId = "virtual:@hiogawa/vite-index-html-middleware/hattip";
+const externalId = "@hiogawa/vite-index-html-middleware/runtime";
+const virtualId = "virtual:" + externalId;
+
 const globalViteDevServerKey =
   "__indexHtmlMiddlewarePlugin_globalViteDevServerKey";
 
 export function indexHtmlMiddlewarePlugin(): Plugin {
   return {
     name: "@hiogawa/vite-index-html-middleware",
+    enforce: "pre", // required to intercept `resolveId`
 
     // expose dev server to access "transformIndexHtml"
     // https://github.com/cyco130/vavite/blob/913e066fd557a1720923361db77c195ac237ac26/packages/expose-vite-dev-server/src/index.ts
@@ -16,13 +18,13 @@ export function indexHtmlMiddlewarePlugin(): Plugin {
       (globalThis as any)[globalViteDevServerKey] = server;
     },
 
-    closeWatcher() {
+    buildEnd() {
       delete (globalThis as any)[globalViteDevServerKey];
     },
 
-    async resolveId(source, _importer, _options) {
-      if (source === virtualId) {
-        return source;
+    async resolveId(source, _importer, options) {
+      if (options.ssr && source === externalId) {
+        return virtualId;
       }
       return;
     },
@@ -31,8 +33,8 @@ export function indexHtmlMiddlewarePlugin(): Plugin {
       // TODO: configurable index.html entry?
       if (id === virtualId) {
         return `
-          import { createIndexHtmlMiddleware } from "@hiogawa/vite-index-html-middleware/runtime";
-          export default createIndexHtmlMiddleware({
+          import { createIndexHtmlMiddleware } from "@hiogawa/vite-index-html-middleware/runtime-internal";
+          export const indexHtmlMiddleware = createIndexHtmlMiddleware({
             server: globalThis[${JSON.stringify(globalViteDevServerKey)}],
             importIndexHtml: () => (import.meta.env.DEV ? import("/index.html?raw") : import("/dist/client/index.html?raw")),
           });
