+++ packages/vite-index-html-middleware/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/runtime.ts"],
+  entry: ["src/index.ts", "src/runtime.ts", "src/runtime-internal.ts"],
   format: ["esm", "cjs"],
   dts: true,
 });
