--- packages/vite-index-html-middleware/virtual.d.ts
@@ -1,4 +1,0 @@
-declare module "virtual:@hiogawa/vite-index-html-middleware/hattip" {
-  const value: import("./dist/runtime").IndexHtmlMiddleware;
-  export default value;
-}
